Align function signature/return type for compose and composeAndValidate
